+++ src/File.h
@@ -12,12 +12,12 @@ public:
 	virtual ~File(void);
 	virtual void OpenOld(const std::string &fileName);
 	virtual void OpenNew(const std::string &fileName);
-	virtual long Length(void);
-	virtual void SetPos(long pos, bool truncate);
-	virtual long GetPos(void);
-	virtual void ReadBytes(long numBytes, void *buffer);
-	virtual void ReadUnicode(long *len, wchar_t **text);
-	virtual void WriteBytes(long numBytes, void *buffer);
+	virtual int Length(void);
+	virtual void SetPos(int pos, bool truncate);
+	virtual int GetPos(void);
+	virtual void ReadBytes(int numBytes, void *buffer);
+	virtual void ReadUnicode(int *len, wchar_t **text);
+	virtual void WriteBytes(int numBytes, void *buffer);
 	virtual void Close(bool truncate);
 private:
 	bool m_error;
